requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
redis==5.0.8
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
